+++ PhpSerializerNET.Test/Deserialize/Validation/TestDoubleValidation.cs
@@ -11,7 +11,7 @@ public class TestDoubleValidation {
 	[Theory]
 	[InlineData("d", "Unexpected end of input. Expected ':' at index 1, but input ends at index 0")]
 	[InlineData("b     ", "Unexpected token at index 1. Expected ':' but found ' ' instead.")]
-	[InlineData("d:111111", "Unexpected end of input. Expected ':' at index 7, but input ends at index 7")]
+	[InlineData("d:111111", "Unexpected end of input. Expected ';' at index 8, but input ends at index 7")]
 	[InlineData("d:bgg5;", "Unexpected token at index 2. 'b' is not a valid part of a floating point number.")]
 	[InlineData("d:;", "Unexpected token at index 2: Expected floating point number, but found ';' instead.")]
 	public void ThrowsOnMalformedDouble(string input, string exceptionMessage) {